Create README.md
This project deals with the increasing traffic problems in cities. We decided to work on this topic due to the following reasons :- Reducing traffic congestion, reducing long time delay, keeps track of vehicles and many more. Also, we have seen that due to traffic emergency vehicles (ambulance,fire brigade) get delayed to reach on time. We will use IR sensors to sense the limit of
the maximum vehicles. The sensors will send the signal to the microcontroller. It will compare which road has the highest density of vehicles and will send the signal to open that particular road for crossing. Thus, reducing the traffic. We are also using sound detecting sensors. Whenever it detects sound in the form of a siren for any emergency vehicle(like ambulance, fire brigade). These sensors will also send the signal to the microcontroller to make that signal green so that the emergency vehicle can pass easily and takes less time. There is also a temperature sensor that can sense the temperature and display it on an LCD display. These two are also connected with microcontroller. The final demonstration of our project will be in a small model form. That side of the road having the highest density of vehicles will glow a green LED. If the sound detector detects any sound that exceeds the threshold value(which is set by us in the code) then also the green LED will glow, of the road where the sensor senses the sound. The LCD will display the temperature sensed by the sensor. It will also display which side of the crossroad has the green signal glowing. So, this is how our project will work.